requirements.txt
cycler==0.10.0
decorator==4.4.2
dlib==19.21.1
imageio==2.9.0
imutils==0.5.4
kiwisolver==1.3.1
matplotlib==3.3.4
networkx==2.5
numpy==1.20.1
opencv-python-headless==4.5.1.48
Pillow==8.1.2
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.18.1
scipy==1.6.1
six==1.15.0
tifffile==2021.3.5
